refactor(desktop): extract repository git settings model
